# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Etsy Mockup Tool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List templates, optionally filtered by product
      parameters:
        - name: productId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a template from a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateBody"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Template:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        name:
          type: string
        type:
          type: string
          description: "M1 or M2"
        optionType:
          type: string
          description: "A or B"
        prompt:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        sessionConfig:
          type: object
          additionalProperties: true
        createdAt:
          type: string
      required:
        - id
        - productId
        - name
        - type
        - optionType
        - prompt
        - imageUrls
        - sessionConfig
        - createdAt
    CreateTemplateBody:
      type: object
      properties:
        sessionId:
          type: string
        name:
          type: string
      required:
        - sessionId
        - name
...